adding new launcher; handling gps from msg
Adding the new rle_kittiplayer.launch : with this one you can use the downloaded datasets and the kitti_player node (called inside the launcher). First you have to download the map using the ira_open_street_map downloader node and creating a dummy launcher with the starting coordinates (not really needed, but the launcher requires this to execute).

"Handling" the gps initialization phase with a variable... fixes #395 @30m