some changes related to refs #443 @1h
the particle score is now:

setParticleScore =
     street_distribution_weight pose_diff_score_component        +
     angle_distribution_weight  final_angle_diff_score_component +
     roadState_distribution_weight getComponentWeight()
then,
     setParticleScore = setParticleScore / 3.0f;

using the maxValue (3) to normalize the particle score into a
probability range.

Moreover, the particles are then multiplied by their previous score.

-----------------------------------------------------------------------

In the resampling phase, the scores are not set to any prefixed value.
bug. even though I remember the scores should be 1/N here I am going to
use the following version:

temp_part.setParticleScore(1.0f);

both in the UNIFORM RESAMPLER and the PROBABILITIC RESAMPLER.

refs #445

-----------------------------------------------------------------------

publishMarkerArray is now using a L-infinity normalizer scheme for
visualizing the particle with the higher score with 0 alpha (brighter).
In this manner I'm able to perceive the *better* particle in RVIZ.