a bug was identified. this commit refs #529
these lines doesn't work

<<roadState->setParticle(&new_particle);>>

this is because the particles are created inside the reconfigureCallBack
routine and then put inside a std:vector with push_back. But vectors creates
copies of objects. The "new particle" I tried to put there as a pointer
inside the component is then deleted once outside the routine itself.
So, to get out of this shit, add that pointer as a new parameter in the
constructor, but I'm not sure it will work either.. it doesn't. Oh fuck..
Need to think about this, for now, I report this issue...