huge list of changes, this commit refs #522 this commit does not close the issue, but now the code compile correctly and it is a middle point towards before refactoring, I'm gonna change what now are colled "weight" but are "alpha-values" instead.
The distances (euclidean and misalignment) are now stored as a new
component called LayoutComponent_OSMDistance; DEPRECATING every
routine/member related to the old behavior. The code was not removed yet
but only commented with a right tag indicating the task 522.

This creates/updates some issues like BUG reported in refs #530 inside
the reconfigureCallback.