merging with ITSC2016 building version
first commit towards the merge with the ITSC_2016 building version.
with this version we should have all the bug fixes of the building
version, i.e.

1. shared pointers corrections
2. building component implementations
3. delta-time fixes
4. tf as pointer, fixed

but without enabling any components (crossing or building). in this
commit there is only the merge with buildings disabled. the previous
behaviour, i.e. LANES components, is still disabled (safe-state).

we should have a version without any component added (ok, only the
OSM component, that is the ITSC2015 version)